moves classifier inference to modules
